增加软删除时，支持设置软删除列为id的值，这样便于软删除的表，建立有业务规则的唯一索引

+++ src/test/resources/create-table.sql
@@ -1,7 +1,7 @@
 DROP TABLE IF EXISTS `t_student`;
 CREATE TABLE `t_student` (
   `id` int(11) NOT NULL AUTO_INCREMENT,
-  `deleted` tinyint(4) DEFAULT NULL,
+  `deleted` int(11) DEFAULT NULL,
   `create_time` datetime DEFAULT NULL,
   `update_time` datetime DEFAULT NULL,
   `delete_time` datetime DEFAULT NULL,
